refactor: five structural cleanups (loader dedup, contract harness split, typed specs, load_data template, base module split) (#26)
* refactor: dedupe fixture loader boilerplate into shared _load_cached pipeline

* refactor: split contract harness into concern mixins and lift row canonicalizer

* refactor: author property-mapping specs through a typed PropertySpec builder

* refactor: hoist load_data prologue into a base template method with _load_rows hook

* refactor: split kg base module into reader_base, readers, and feature_group with a re-export facade

* fix: materialize one-shot allowed_values iterables in PropertySpec
