+++ open_kgo/feature_groups/kg/agent_memory/base.py
@@ -4,14 +4,13 @@ from __future__ import annotations
 
 from typing import Any, ClassVar
 
-from mloda.core.abstract_plugins.components.default_options_key import DefaultOptionKeys
-
 from open_kgo.feature_groups.kg.base import (
     KgConnectorFeatureGroupBase,
     QueryReader,
     compose_property_mapping,
 )
 from open_kgo.feature_groups.kg.mixins import PaginationMixin
+from open_kgo.feature_groups.kg.spec import property_spec
 
 
 _RETRIEVAL_MODES: dict[str, str] = {
@@ -49,13 +48,7 @@ _MEMORY_SCOPE_SPECS: tuple[tuple[str, str, None | tuple[()]], ...] = (
 MEMORY_SCOPE_KEYS: tuple[str, ...] = tuple(name for name, _, _ in _MEMORY_SCOPE_SPECS)
 
 _MEMORY_SCOPE_PROPERTY_MAPPING: dict[str, Any] = {
-    name: {
-        "explanation": explanation,
-        DefaultOptionKeys.context: True,
-        DefaultOptionKeys.strict_validation: False,
-        DefaultOptionKeys.default: default,
-    }
-    for name, explanation, default in _MEMORY_SCOPE_SPECS
+    name: property_spec(explanation, default=default) for name, explanation, default in _MEMORY_SCOPE_SPECS
 }
 
 
@@ -65,43 +58,31 @@ class AgentMemoryReader(PaginationMixin, QueryReader):
         PaginationMixin.PROPERTY_MAPPING_DELTA,
         _MEMORY_SCOPE_PROPERTY_MAPPING,
         {
-            "reference_time": {
-                "explanation": "Bi-temporal reference time (ISO 8601).",
-                DefaultOptionKeys.context: True,
-                DefaultOptionKeys.strict_validation: False,
-                DefaultOptionKeys.default: None,
-            },
-            "valid_at_range": {
-                "explanation": "[start, end] for valid_at filter.",
-                DefaultOptionKeys.context: True,
-                DefaultOptionKeys.strict_validation: False,
-                DefaultOptionKeys.default: (),
-            },
-            "invalid_at_range": {
-                "explanation": "[start, end] for invalid_at filter.",
-                DefaultOptionKeys.context: True,
-                DefaultOptionKeys.strict_validation: False,
-                DefaultOptionKeys.default: (),
-            },
-            "retrieval_mode": {
-                "explanation": "Retrieval strategy used to score candidate memories.",
-                "allowed_values": _RETRIEVAL_MODES,
-                DefaultOptionKeys.context: True,
-                DefaultOptionKeys.strict_validation: True,
-                DefaultOptionKeys.default: "lexical",
-            },
-            "mmr_lambda": {
-                "explanation": "MMR lambda for hybrid retrieval blend (0.0-1.0).",
-                DefaultOptionKeys.context: True,
-                DefaultOptionKeys.strict_validation: False,
-                DefaultOptionKeys.default: 0.5,
-            },
-            "threshold": {
-                "explanation": "Similarity threshold (0.0-1.0).",
-                DefaultOptionKeys.context: True,
-                DefaultOptionKeys.strict_validation: False,
-                DefaultOptionKeys.default: 0.0,
-            },
+            "reference_time": property_spec(
+                "Bi-temporal reference time (ISO 8601).",
+            ),
+            "valid_at_range": property_spec(
+                "[start, end] for valid_at filter.",
+                default=(),
+            ),
+            "invalid_at_range": property_spec(
+                "[start, end] for invalid_at filter.",
+                default=(),
+            ),
+            "retrieval_mode": property_spec(
+                "Retrieval strategy used to score candidate memories.",
+                strict=True,
+                allowed_values=_RETRIEVAL_MODES,
+                default="lexical",
+            ),
+            "mmr_lambda": property_spec(
+                "MMR lambda for hybrid retrieval blend (0.0-1.0).",
+                default=0.5,
+            ),
+            "threshold": property_spec(
+                "Similarity threshold (0.0-1.0).",
+                default=0.0,
+            ),
         },
         context="AgentMemoryReader",
     )
