+++ open_kgo/feature_groups/kg/agent_memory/graph_walk_memory.py
@@ -39,6 +39,7 @@ from open_kgo.feature_groups.kg.agent_memory.base import (
     AgentMemoryFeatureGroup,
     AgentMemoryReader,
 )
+from open_kgo.feature_groups.kg.base import LoadContext
 from open_kgo.feature_groups.kg.errors import FixtureLoadError, UnknownMemoryScopeError
 from open_kgo.feature_groups.kg.fixtures import load_json_fixture
 
@@ -151,9 +152,8 @@ class GraphWalkMemoryReader(AgentMemoryReader):
         return text.strip()
 
     @classmethod
-    def load_data(cls, data_access: Any, features: FeatureSet) -> list[dict[str, Any]]:
-        ctx = cls._prepare_load(data_access)
-        graph = cls._connect_from_slot(ctx.slot)
+    def _load_rows(cls, ctx: LoadContext, connection: Any, features: FeatureSet) -> list[dict[str, Any]]:
+        graph = connection
         seed = cls.build_query(features)
 
         if seed not in graph:
