+++ open_kgo/feature_groups/kg/citation_rest/base.py
@@ -4,35 +4,26 @@ from __future__ import annotations
 
 from typing import Any, ClassVar
 
-from mloda.core.abstract_plugins.components.default_options_key import DefaultOptionKeys
-
 from open_kgo.feature_groups.kg.base import (
     KgConnectorFeatureGroupBase,
     ParamReader,
     compose_property_mapping,
 )
 from open_kgo.feature_groups.kg.mixins import PaginationMixin
+from open_kgo.feature_groups.kg.spec import property_spec
 
 
 _PER_CALL_KEYS: dict[str, Any] = {
-    "entity_type": {
-        "explanation": "Resource type (e.g. 'pathway', 'work').",
-        DefaultOptionKeys.context: True,
-        DefaultOptionKeys.strict_validation: False,
-        DefaultOptionKeys.default: None,
-    },
-    "stable_id": {
-        "explanation": "System-stable identifier of the entity to fetch (e.g. R-HSA-1640170).",
-        DefaultOptionKeys.context: True,
-        DefaultOptionKeys.strict_validation: False,
-        DefaultOptionKeys.default: None,
-    },
-    "hierarchy_depth": {
-        "explanation": "Depth limit for ancestors/descendants traversal.",
-        DefaultOptionKeys.context: True,
-        DefaultOptionKeys.strict_validation: False,
-        DefaultOptionKeys.default: 1,
-    },
+    "entity_type": property_spec(
+        "Resource type (e.g. 'pathway', 'work').",
+    ),
+    "stable_id": property_spec(
+        "System-stable identifier of the entity to fetch (e.g. R-HSA-1640170).",
+    ),
+    "hierarchy_depth": property_spec(
+        "Depth limit for ancestors/descendants traversal.",
+        default=1,
+    ),
 }
 
 
@@ -41,18 +32,12 @@ class CitationRestReader(PaginationMixin, ParamReader):
         ParamReader.PROPERTY_MAPPING,
         PaginationMixin.PROPERTY_MAPPING_DELTA,
         {
-            "species_prefix": {
-                "explanation": "Species prefix (e.g. HSA for human Reactome IDs).",
-                DefaultOptionKeys.context: True,
-                DefaultOptionKeys.strict_validation: False,
-                DefaultOptionKeys.default: None,
-            },
-            "dataset_version": {
-                "explanation": "Release version pin (e.g. 'v90' for Reactome).",
-                DefaultOptionKeys.context: True,
-                DefaultOptionKeys.strict_validation: False,
-                DefaultOptionKeys.default: None,
-            },
+            "species_prefix": property_spec(
+                "Species prefix (e.g. HSA for human Reactome IDs).",
+            ),
+            "dataset_version": property_spec(
+                "Release version pin (e.g. 'v90' for Reactome).",
+            ),
         },
         context="CitationRestReader",
     )
