+++ open_kgo/feature_groups/kg/embedded/networkx_embedded.py
@@ -21,6 +21,7 @@ import networkx as nx
 
 from mloda.core.abstract_plugins.components.feature_set import FeatureSet
 
+from open_kgo.feature_groups.kg.base import LoadContext
 from open_kgo.feature_groups.kg.embedded.base import (
     EmbeddedGraphFeatureGroup,
     EmbeddedGraphReader,
@@ -76,9 +77,8 @@ class NetworkxEmbeddedReader(EmbeddedGraphReader):
         return loader(locator)
 
     @classmethod
-    def load_data(cls, data_access: Any, features: FeatureSet) -> list[dict[str, Any]]:
-        ctx = cls._prepare_load(data_access)
-        graph = cls._connect_from_slot(ctx.slot)
+    def _load_rows(cls, ctx: LoadContext, connection: Any, features: FeatureSet) -> list[dict[str, Any]]:
+        graph = connection
 
         params = cls.build_params(features, ctx.slot)
         op = params["operation"]
