+++ open_kgo/feature_groups/kg/lineage/dbt_manifest.py
@@ -12,6 +12,7 @@ from typing import Any, ClassVar, Mapping
 
 from mloda.core.abstract_plugins.components.feature_set import FeatureSet
 
+from open_kgo.feature_groups.kg.base import LoadContext
 from open_kgo.feature_groups.kg.fixtures import copy_cached_row, load_json_fixture
 from open_kgo.feature_groups.kg.lineage.base import LineageFeatureGroup, LineageReader
 
@@ -44,9 +45,8 @@ class DbtManifestReader(LineageReader):
         return load_json_fixture(cls.CONNECTOR_ID, slot["locator"])
 
     @classmethod
-    def load_data(cls, data_access: Any, features: FeatureSet) -> list[dict[str, Any]]:
-        ctx = cls._prepare_load(data_access)
-        manifest = cls._connect_from_slot(ctx.slot)
+    def _load_rows(cls, ctx: LoadContext, connection: Any, features: FeatureSet) -> list[dict[str, Any]]:
+        manifest = connection
         params = cls.build_params(features, ctx.slot)
 
         asset_urn = params.get("asset_urn")
