+++ open_kgo/feature_groups/kg/rdf/rdflib_sparql.py
@@ -18,6 +18,7 @@ import rdflib
 
 from mloda.core.abstract_plugins.components.feature_set import FeatureSet
 
+from open_kgo.feature_groups.kg.base import LoadContext
 from open_kgo.feature_groups.kg.fixtures import _rejected_scheme, load_rdf_graph
 from open_kgo.feature_groups.kg.rdf.base import RdfSparqlFeatureGroup, RdfSparqlReader
 
@@ -76,10 +77,9 @@ class RdfLibSparqlReader(RdfSparqlReader):
         return load_rdf_graph(cls.CONNECTOR_ID, locator)
 
     @classmethod
-    def load_data(cls, data_access: Any, features: FeatureSet) -> list[dict[str, Any]]:
+    def _load_rows(cls, ctx: LoadContext, connection: Any, features: FeatureSet) -> list[dict[str, Any]]:
         """Run the SPARQL query and return up to result_limit rows as list-of-dicts."""
-        ctx = cls._prepare_load(data_access)
-        graph = cls._connect_from_slot(ctx.slot)
+        graph = connection
 
         query_text = cls.build_query(features)
         rows: list[dict[str, Any]] = []
