+++ open_kgo/feature_groups/kg/rest_public/file_fixture_paged_rest.py
@@ -47,6 +47,7 @@ from typing import Any, ClassVar, Mapping
 
 from mloda.core.abstract_plugins.components.feature_set import FeatureSet
 
+from open_kgo.feature_groups.kg.base import LoadContext
 from open_kgo.feature_groups.kg.fixtures import copy_cached_row, load_json_fixture
 from open_kgo.feature_groups.kg.mixins import parse_page_size
 from open_kgo.feature_groups.kg.rest_public.base import (
@@ -100,9 +101,8 @@ class FileFixturePagedRestReader(RestPublicReader):
         return path
 
     @classmethod
-    def load_data(cls, data_access: Any, features: FeatureSet) -> list[dict[str, Any]]:
-        ctx = cls._prepare_load(data_access)
-        path = cls._connect_from_slot(ctx.slot)
+    def _load_rows(cls, ctx: LoadContext, connection: Any, features: FeatureSet) -> list[dict[str, Any]]:
+        path = connection
 
         page_size = parse_page_size(cls.CONNECTOR_ID, ctx.slot.get("page_size"), 100)
         pages_dir = path if path.is_dir() else path.parent
