+++ open_kgo/feature_groups/kg/rest_public/file_fixture_rest.py
@@ -32,7 +32,7 @@ from typing import Any, ClassVar, Mapping
 
 from mloda.core.abstract_plugins.components.feature_set import FeatureSet
 
-from open_kgo.feature_groups.kg.base import narrow_property_mapping
+from open_kgo.feature_groups.kg.base import LoadContext, narrow_property_mapping
 from open_kgo.feature_groups.kg.fixtures import copy_cached_row, load_json_fixture
 from open_kgo.feature_groups.kg.rest_public.base import (
     RestPublicFeatureGroup,
@@ -81,9 +81,8 @@ class FileFixtureRestReader(RestPublicReader):
         return path
 
     @classmethod
-    def load_data(cls, data_access: Any, features: FeatureSet) -> list[dict[str, Any]]:
-        ctx = cls._prepare_load(data_access)
-        path = cls._connect_from_slot(ctx.slot)
+    def _load_rows(cls, ctx: LoadContext, connection: Any, features: FeatureSet) -> list[dict[str, Any]]:
+        path = connection
 
         pages_dir = path if path.is_dir() else path.parent
         # Sort numerically on the ``<N>`` in ``page_<N>.json``: a lexical
